requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
chromadb
groq
sentence-transformers